sasg_code/requirements.txt
ecdsa==0.16.1
Flask==1.1.2
Flask-Bootstrap==3.3.6.0
Flask-LDAP==0.1.6
Flask-Login==0.3.2
Flask-Migrate==1.8.0
Flask-OpenID==1.2.5
Flask-PyMongo==0.4.1
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.3
ipaddress==1.0.16
junos-eznc==2.2.1
paramiko==2.7.1
pycryptodome==3.10.1
sphinx_rtd_theme==0.4.3
sphinx==1.8.4
uWSGI==2.0.17
werkzeug==0.16.1
WTForms==2.1
xmlplain==1.6.0